[pvr.tvh] implement GetPlayingTime()
What it means is we convert tvheadend's notion of "seconds shifted" to an
absolute timestamp.

This fixes two issues. The first one is that if you pause a live stream at e.g.
10:00 program time (ie. 10 minutes after it has started), wait for 30 seconds,
unpause and then rewind 30 seconds, XBMC would display the current time as
11:00 when in fact it should be 10:00 again (since we rewinded the buffer to
the beginning).

The other issue is more of a new feature. Now when you pause a live stream
long enough XBMC will display the shifted position along with the live position
in the seek bar, so you know roughly where in the shift you are.